+++ packages/check/src/main.rs
@@ -4,7 +4,7 @@ use std::io::Read;
 use std::path::Path;
 use std::process::exit;
 
-use clap::{App, Arg};
+use clap::{Arg, ArgAction, Command};
 use colored::Colorize;
 
 use cosmwasm_vm::capabilities_from_csv;
@@ -13,38 +13,43 @@ use cosmwasm_vm::internals::{check_wasm, compile};
 const DEFAULT_AVAILABLE_CAPABILITIES: &str = "iterator,staking,stargate,cosmwasm_1_1,cosmwasm_1_2";
 
 pub fn main() {
-    let matches = App::new("Contract checking")
+    let matches = Command::new("Contract checking")
         .version(env!("CARGO_PKG_VERSION"))
         .long_about("Checks the given wasm file (memories, exports, imports, available capabilities, and non-determinism).")
         .author("Mauro Lacy <mauro@lacy.com.es>")
         .arg(
-            Arg::with_name("CAPABILITIES")
+            Arg::new("CAPABILITIES")
                 // `long` setting required to turn the position argument into an option 🤷
                 .long("available-capabilities")
                 .aliases(&["FEATURES", "supported-features"]) // Old names
                 .value_name("CAPABILITIES")
                 .help("Sets the available capabilities that the desired target chain has")
-                .takes_value(true)
+                .num_args(1)
+                .action(ArgAction::Set),
         )
         .arg(
-            Arg::with_name("WASM")
+            Arg::new("WASM")
                 .help("Wasm file to read and compile")
                 .required(true)
                 .index(1)
-                .multiple(true),
+                .num_args(0..)
+                .action(ArgAction::Append),
         )
         .get_matches();
 
     // Available capabilities
     let available_capabilities_csv = matches
-        .value_of("CAPABILITIES")
+        .get_one::<String>("CAPABILITIES")
+        .map(|s| s.as_str())
         .unwrap_or(DEFAULT_AVAILABLE_CAPABILITIES);
     let available_capabilities = capabilities_from_csv(available_capabilities_csv);
     println!("Available capabilities: {:?}", available_capabilities);
     println!();
 
     // File
-    let paths = matches.values_of("WASM").expect("Error parsing file names");
+    let paths = matches
+        .get_many::<String>("WASM")
+        .expect("Error parsing file names");
 
     let (passes, failures): (Vec<_>, _) = paths
         .map(|p| {